requirements.txt
astroid==1.5.3
autopep8==1.3.2
colorama==0.3.9
coverage==4.4.1
Django==1.11.4
django-extensions==1.8.1
django-filter==1.0.4
djangorestframework==3.6.3
factory-boy==2.9.2
Faker==0.7.18
flake8==3.4.1
flake8-polyfill==1.0.1
isort==4.2.15
lazy-object-proxy==1.3.1
mando==0.6.4
mccabe==0.6.1
numpy==1.13.1
Pillow==4.2.1
pycodestyle==2.3.1
pyflakes==1.5.0
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2017.2
radon==2.0.2
six==1.10.0
wrapt==1.10.10
djangorestframework-jwt==1.11.0
django-rest-swagger==2.1.2
django-role-permissions==2.1.0
pyyaml==3.12
